third_party/WebKit/LayoutTests/editing/inserting/insert-div-020.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '<span id="test">foo</span><br><br>|bar',
        '</div>',
    ].join(''),
    'insertParagraph',
    [
        '<div contenteditable>',
            '<span id="test">foo</span><br><br><br><div>|bar</div>',
        '</div>',
    ].join('')),
    'Insert BR and enclose "bar" with DIV');
</script>
